Cleanup the MCP bridge
- Cache McpClient in ForwardClient and share it across bridges
- Make bridges address-independent and cache ForwardClient entries in registry
- Replace ForwardRegistry interface with concrete bridge registry
- Use McpToolsBridge and McpResourceBridge directly in ForwardsBean
- Extract MCP client transport logic into McpBridge and DefaultMcpBridge
